refactor: 딜러와 플레이어 간 중복되는 코드가 없도록 abstract 클래스를 상속받도록 수정
