+++ src/main/kotlin/blackjack/model/Participant.kt
@@ -1,3 +1,18 @@
 package blackjack.model
 
-data class Participant(val dealer: Dealer, val players: List<Player>)
+abstract class Participant(val hand: Hand) {
+    fun hit(card: Card) {
+        hand.add(card)
+    }
+
+    fun hitIfConditionTrue(
+        deck: Deck,
+        condition: () -> Boolean,
+        view: () -> Unit,
+    ) {
+        while (condition()) {
+            hit(deck.pull())
+            view()
+        }
+    }
+}
