+++ src/main/kotlin/blackjack/model/ParticipantData.kt
@@ -1,0 +1,3 @@
+package blackjack.model
+
+data class ParticipantData(val dealer: Dealer, val players: List<Player>)
